Two screwed things solved: 1 - the jcr provider was removing the indexes at startup. Even if it wasn't temporary 2 - for some strange reason I excluded the postgres-repository.xml before. It came back.

+++ osl-jcr/src/main/java/org/openspotlight/jcr/provider/JcrConnectionProvider.java
@@ -150,11 +150,14 @@ public abstract class JcrConnectionProvider {
 		public Repository internalOpenRepository() {
 			if (repository == null || repositoryClosed) {
 				try {
-					try {
-						Files.delete(getData().getConfigurationDirectory());
-					} catch (final SLException e) {
-						throw Exceptions.logAndReturnNew(e,
-								SLRuntimeException.class);
+					if (getData().isTemporary()) {
+						try {
+							Files.delete(getData().getConfigurationDirectory());
+						} catch (final SLException e) {
+							throw Exceptions.logAndReturnNew(e,
+									SLRuntimeException.class);
+
+						}
 					}
 
 					final RepositoryConfig config = RepositoryConfig.create(
